providers/gcp/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
google-cloud-storage>=2.10.0
urllib3>=2.0.0
